fix: fix big transfer test
