+++ jinahub/encoders/image/BigTransferEncoder/config.yml
@@ -1,0 +1,6 @@
+jtype: BigTransferEncoder
+with:
+  on_gpu: true
+metas:
+  py_modules:
+    - big_transfer.py
