+++ jinahub/encoders/image/BigTransferEncoder/requirements.txt
@@ -1,0 +1,4 @@
+jina-commons @ git+https://github.com/jina-ai/jina-commons.git#egg=jina-commons
+tensorflow>=2.5.0
+numpy
+Pillow
