+++ jinahub/encoders/image/BigTransferEncoder/tests/requirements.txt
@@ -1,0 +1,2 @@
+jina==2.0.8
+pytest==6.2.4
