--- jinahub/encoders/text/BigTransferEncoder/.dockerignore
@@ -1,24 +1,0 @@
-pretrained
-venv
-.venv
-.github
-__pycache__/
-build
-dist
-*.egg-info/
-.idea
-tests
-
-# Unit test / coverage reports
-htmlcov/
-.tox/
-.nox/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-*.py,cover
-.hypothesis/
-.pytest_cache/
