--- jinahub/encoders/text/BigTransferEncoder/Dockerfile.gpu
@@ -1,10 +1,0 @@
-FROM nvidia/cuda:11.3.1-cudnn8-runtime
-
-RUN apt update && apt install -y python3.8 python3-pip
-
-COPY . /big_transfer/
-WORKDIR /big_transfer
-
-RUN pip install .
-
-ENTRYPOINT ["jina", "executor", "--uses", "config.yml"]
