--- jinahub/encoders/text/BigTransferEncoder/config.yml
@@ -1,6 +1,0 @@
-jtype: BigTransferEncoder
-with:
-  on_gpu: true
-metas:
-  py_modules:
-    - big_transfer.py
