--- jinahub/encoders/text/BigTransferEncoder/manifest.yml
@@ -1,8 +1,0 @@
-manifest_version: 1
-name: BigTransferEncoder
-alias: BigTransferEncoder
-description: '`BigTransferEncoder` encodes images using a Big Transfer Model'
-author: Jina AI Dev-Team (dev-team@jina.ai)
-url: https://github.com/jina-ai/executor-big-transfer-encoder
-avatar: https://avatars.githubusercontent.com/u/60539444?s=200&v=4
-keywords: [image, encoder, Tensorflow, BiT, big_transfer, resnet]
