--- jinahub/encoders/text/BigTransferEncoder/tests/integration/flow.yml
@@ -1,12 +1,0 @@
-jtype: Flow
-version: '1'
-with:
-  return_results: true
-  workspace: $WORKSPACE
-executors:
-  - name: encoder
-    uses:
-      jtype: BigTransferEncoder
-      with:
-        model_name: $TRANSFER_MODEL_NAME
-        on_gpu: true
