--- jinahub/encoders/text/BigTransferEncoder/tests/requirements.txt
@@ -1,2 +1,0 @@
-jina==2.0.8
-pytest==6.2.4
